use roaring bitmaps (#15)
* new ideas but mixed with ai madness

* do some funky refactoring

* try the new set instead of maps

* make stock simpler

+++ pkg/facet/types.go
@@ -9,7 +9,7 @@ type FieldResult interface {
 }
 
 type KeyFieldResult struct {
-	Values map[string]int `json:"values,omitempty"`
+	Values map[string]uint64 `json:"values,omitempty"`
 }
 
 func (k *KeyFieldResult) HasValues() bool {
